backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-pipeline"
version = "0.1.0"
description = "Weather data pipeline for ski resort forecasts"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Nate", email = "nate@example.com"}
]

dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "duckdb>=0.9.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
    "streamlit>=1.28.0",
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

prefect = [
    "prefect>=2.14.0",
]

all = [
    "weather-pipeline[dev,prefect]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "clients*", "db*", "flows*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by black)

[project.scripts]
init-db = "db.session:init_db"
